% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buthesis.R
\docType{package}
\name{buthesis}
\alias{buthesis}
\alias{buthesis-package}
\title{Baylor University Department of Statistical Science Thesis Templates}
\description{
This package contains latex and knitr templates for the statistical science department
that adhere to the Baylor Graduate School guidlines. The templates are located in the \code{inst/} directory.
}
\details{
buthesis: Thesis templates for the Statistical Science Department at Baylor University
}
