% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_templates.R
\name{create_latex_thesis}
\alias{create_latex_thesis}
\title{Create a directory that contains the proper Baylor Statistical Science thesis latex template}
\usage{
create_latex_thesis(direc)
}
\arguments{
\item{direc}{path to a directory where the template will reside. If the directory does not exist, the function will create it.}
}
\value{
nothing
}
\description{
This function creates a specified directory and populates the directory with the latex template.
}
